+++ core/datastore/src/main/kotlin/io/github/vedicmitra/core/datastore/di/DataStoreModule.kt
@@ -23,9 +23,11 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import io.github.vedicmitra.core.datastore.DefaultLocationRepository
+import io.github.vedicmitra.core.datastore.DefaultProfileRepository
 import io.github.vedicmitra.core.datastore.DefaultReminderRepository
 import io.github.vedicmitra.core.datastore.DefaultUserPreferencesRepository
 import io.github.vedicmitra.core.datastore.LocationRepository
+import io.github.vedicmitra.core.datastore.ProfileRepository
 import io.github.vedicmitra.core.datastore.ReminderRepository
 import io.github.vedicmitra.core.datastore.UserPreferencesRepository
 import javax.inject.Singleton
@@ -50,6 +52,9 @@ internal abstract class DataStoreModule {
 
     @Binds
     abstract fun bindLocationRepository(impl: DefaultLocationRepository): LocationRepository
+
+    @Binds
+    abstract fun bindProfileRepository(impl: DefaultProfileRepository): ProfileRepository
 
     companion object {
         @Provides